+++ model/src/main/java/org/kethereum/model/Transaction.kt
@@ -9,7 +9,7 @@ data class Transaction(val value: BigInteger,
                        val to: Address?,
 
                        val creationEpochSecond: Long? = null,
-                       var nonce: Long? = null,
+                       var nonce: BigInteger? = null,
                        var gasPrice: BigInteger = DEFAULT_GAS_PRICE,
                        var gasLimit: BigInteger = DEFAULT_GAS_LIMIT,
                        var sigHash: String? = null,
